# Conflicts:
#	examples/CMakeLists.txt
#	examples/tm_ultraface.cpp

+++ src/dev/cpu/op/conv/x86/conv_dw_kernel_x86.c
@@ -113,7 +113,7 @@ static void convdw3x3s1(float* output, float* img_data, float* kernel_data, floa
     int channel_count = inc >> 3;
     int channel_remain = inc - (channel_count << 3);
     // generate the image tmp
-    float* img_tmp = ( float* )sys_malloc(8 * inwh * (channel_count + 1) * sizeof(float));
+    float* img_tmp = ( float* )sys_malloc(8 * (unsigned long)inwh * (channel_count + 1) * sizeof(float));
     float* kernel_tmp = ( float* )sys_malloc(8 * 9 * (channel_count + 1) * sizeof(float));
     float* bias_tmp = ( float* )sys_malloc(8 * (channel_count + 1) * sizeof(float));
     {
@@ -320,7 +320,7 @@ static void convdw3x3s1(float* output, float* img_data, float* kernel_data, floa
         }
     }
 
-    float* output_tmp = ( float* )sys_malloc(outwh * (channel_count + 1) * 8 * sizeof(float));
+    float* output_tmp = ( float* )sys_malloc((unsigned long)outwh * (channel_count + 1) * 8 * sizeof(float));
     for (int c = 0; c < channel_count + 1; c++)
     {
         float* ktmp = kernel_tmp + c * 8 * 9;
@@ -769,7 +769,7 @@ static void convdw3x3s2(float* output, float* img_data, float* kernel_data, floa
     int channel_count = inc >> 3;
     int channel_remain = inc - (channel_count << 3);
     // generate the image tmp
-    float* img_tmp = ( float* )sys_malloc(8 * inwh * (channel_count + 1) * sizeof(float));
+    float* img_tmp = ( float* )sys_malloc(8 * (unsigned long)inwh * (channel_count + 1) * sizeof(float));
     float* kernel_tmp = ( float* )sys_malloc(8 * 9 * (channel_count + 1) * sizeof(float));
     float* bias_tmp = ( float* )sys_malloc(8 * (channel_count + 1) * sizeof(float));
     {
@@ -979,7 +979,7 @@ static void convdw3x3s2(float* output, float* img_data, float* kernel_data, floa
         }
     }
 
-    float* output_tmp = ( float* )sys_malloc(outwh * (channel_count + 1) * 8 * sizeof(float));
+    float* output_tmp = ( float* )sys_malloc((unsigned long)outwh * (channel_count + 1) * 8 * sizeof(float));
     for (int c = 0; c < channel_count + 1; c++)
     {
         float* ktmp = kernel_tmp + c * 8 * 9;
@@ -2551,7 +2551,7 @@ int conv_dw_run(struct ir_tensor* input_tensor, struct ir_tensor* weight_tensor,
         input_tmp = input;
     else
     {
-        input_tmp = ( float* )sys_malloc(inh_tmp * inw_tmp * group * sizeof(float));
+        input_tmp = ( float* )sys_malloc((unsigned long)inh_tmp * inw_tmp * group * sizeof(float));
 #pragma omp parallel for num_threads(num_thread)        
         for (int g = 0; g < group; g++)
         {
